assignment2_2024/requirements.txt
haversine==2.8.1
mysql-connector-python==8.0.33
tabulate==0.9.0
python-dotenv==1.0.0
